自機回転をイージング関数で実装できた。原因は関数の引数の型が違ったことで切り捨てされていたこと。

+++ DirectXGame/Engin/3d/Player.h
@@ -65,7 +65,7 @@ private:
 	bool isPush;
 	int pushTime;
 	bool isAttack;
-	int attackTime;
+	float attackTime;
 	//fever
 	int feverTime;
 	bool isFever;
